+++ mycodo/actions/display_flash_on.py
@@ -13,7 +13,7 @@ ACTION_INFORMATION = {
     'name_unique': 'display_flash_on',
     'name': f"{TRANSLATIONS['display']['title']}: {lazy_gettext('Flashing')}: {lazy_gettext('On')}",
     'library': None,
-    'manufacturer': 'Display',
+    'manufacturer': 'Mycodo',
     'application': ['functions'],
 
     'url_manufacturer': None,